Added dtype to storing result data

+++ src/training.py
@@ -45,12 +45,15 @@ def train_ensembles(data, ensembles, l_name, global_epochs=10, convolutional=Fal
     if l_name.startswith("lof"):
         local_scores = - np.array([model.negative_outlier_factor_ for i, model in enumerate(local_detectors)])
     if l_name == "xstream":
-        local_scores = np.array([-model.score(fdata[i]) for i, model in enumerate(local_detectors)])
+        local_scores = np.array([-model.score(fdata[i]) for i, model in enumerate(local_detectors)],
+                                dtype=float)
     if l_name == "if":
-        local_scores = -np.array([model.score_samples(fdata[i]) for i, model in enumerate(local_detectors)])
+        local_scores = -np.array([model.score_samples(fdata[i]) for i, model in enumerate(local_detectors)],
+                                 dtype=float)
     if l_name == "ae":
         if convolutional:
-            predicted = np.array([model.predict(data[i]) for i, model in enumerate(local_detectors)])
+            predicted = np.array([model.predict(data[i]) for i, model in enumerate(local_detectors)],
+                                 dtype=float)
             predicted = predicted.reshape(fshape)
         else:
             print(fdata.shape)
